requirements.txt
# fastapi
# uvicorn
# pyserial
# pandas

fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pyserial>=3.5
requests>=2.31.0
pydantic>=2.0.0
python-multipart>=0.0.9